More improvements and final cleanup
Removed ALL unnecessary @JvmField and @JvmStatic
- it will break some scripts, but they were already broken with b77 RotationUtils port
TODO: utils could be provided with new api objects?

* Refactored `MovementUtils.isMoving`, `ClientUtils.displayChatMessage` and such to just `isMoving`, ... across all modules.

MovementUtils:
* Improved MovementUtils.speed: you can now do speed *= 1.08 and it will essentially call strafe(speed * 1.08) -> simplifies some Speed modules

ColorUtils:
* Implemented optional parameters into rainbow function.

BungeeCordSpoof:
* converted to object class

Sprint:
* fixed horizontal and vertical values isSupported override for newly introduced Legit mode

FriendsConfig:
* simplified serialization, deserialization
* removed old config support

New ClickGUI Panels:
* only move panels to windows border when mouse button is held
  -> prevents them from being moved to center when ClickGUI is opened in smaller window

Code improvement:
* Replaced PrintWriters to Kotlin's file.writeText.
* Replaced .equals to just == across all modules.
  -> simplified some if else blocks to when blocks
* Replaced Random().nextInt and such to just nextInt() from RandomUtils, Random.nextBoolean() -> nextBoolean(), Math.random() * ... -> nextFloat(min, max) ...
  + made parameters for RandomUtils functions optional
* Simplified some function blocks `{}` to just `= ...`
* Removed some unnecessary type annotations, null-safe types

MSTimer:
* hasTimePassed: Removed unnecessary casts to Long across all modules

ProphuntESP:
* fixed box, otherbox modes?
* impossible condition `(!mode.equals("Box", true) || !mode.equals("OtherBox", true))` -> && instead of ||

Updated Kotlin to 1.8.10

+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/player/AntiAFK.kt
@@ -108,13 +108,13 @@ class AntiAFK : Module() {
                 if (jumpValue.get() && thePlayer.onGround)
                     thePlayer.jump()
 
-                if (rotateValue.get() && delayTimer.hasTimePassed(rotationDelayValue.get().toLong())) {
+                if (rotateValue.get() && delayTimer.hasTimePassed(rotationDelayValue.get())) {
                     thePlayer.fixedSensitivityYaw += rotationAngleValue.get()
                     thePlayer.fixedSensitivityPitch += RandomUtils.nextFloat(0F, 1F) * 2 - 1
                     delayTimer.reset()
                 }
 
-                if (swingValue.get() && !thePlayer.isSwingInProgress && swingDelayTimer.hasTimePassed(swingDelayValue.get().toLong())) {
+                if (swingValue.get() && !thePlayer.isSwingInProgress && swingDelayTimer.hasTimePassed(swingDelayValue.get())) {
                     thePlayer.swingItem()
                     swingDelayTimer.reset()
                 }
